Fixing bugs

+++ src/ShopSite/ViewModels/ProductsByCategoryVM.cs
@@ -11,8 +11,8 @@ namespace ShopSite.ViewModels
 
         public int TotalProducts { get; set; }
 
-        public int? SearchMaxPrice { get; set; }
-        public int? SearchMinPrice { get; set; }
+        public decimal? SearchMaxPrice { get; set; }
+        public decimal? SearchMinPrice { get; set; }
 
         public decimal MaxPrice { get; set; }
         public decimal MinPrice { get; set; }
